Adding a helper class for common functionalities
